Get tests passing
