+++ src/vivarium_public_health/dataset_manager/__init__.py
@@ -1,0 +1,2 @@
+from .artifact import Artifact, EntityKey, ArtifactException
+from .dataset_manager import ArtifactManager, ArtifactManagerInterface, get_location_term
