+++ tests/test_dataset_manager.py
@@ -14,7 +14,7 @@ def test_artifact_load_empty_data(mocker):
     uncached_load_mock = mocker.patch('vivarium_public_health.dataset_manager.Artifact._uncached_load')
     uncached_load_mock.return_value = None
 
-    test_artifact = Artifact('test_path', 2010, 2012, 0, 'test_location')
+    test_artifact = Artifact('test_path')
     with pytest.raises(Exception) as e:
         test_artifact.load('test_entity_path')
     assert e.typename == 'ArtifactException'
